- Thêm lớp trung gian MovingObject giữa lớp Mario và GameObject với tham số mới vy (vận tốc theo trục y). - Tách lớp Brick ra thành lớp riêng.
