+++ 02-Sprite/Brick.cpp
@@ -1,0 +1,11 @@
+#include "Brick.h"
+#include "Animations.h"
+
+
+void CBrick::Render() {
+
+	LPANIMATION ani = CAnimations::GetInstance()->Get(510);
+
+	ani->Render(x, y);
+
+}
