+++ 02-Sprite/Brick.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "GameObject.h"
+
+class CBrick : public CGameObject {
+public:
+	CBrick(float x, float y) : CGameObject(x, y) {}
+	void Render();
+	void Update(DWORD dt) {}
+};
