+++ 02-Sprite/MovingObject.cpp
@@ -1,0 +1,2 @@
+#include "MovingObject.h"
+
