+++ samples.sln
@@ -1,19 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31515.178
+# Visual Studio Version 17
+VisualStudioVersion = 17.13.35825.156 d17.13
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "01-Skeleton", "01-Skeleton\01-Skeleton.vcxproj", "{959C78E1-3BC0-4E30-BB42-048A1F251A85}"
-EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "02-Sprite", "02-Sprite\02-Sprite.vcxproj", "{AADFB8E6-BDA3-4F69-9F0F-5C30EE6EACBA}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "03-Keyboard-States", "03-Keyboard-States\03-Keyboard-States.vcxproj", "{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "04-Collision", "04-Collision\04-Collision.vcxproj", "{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "00-Intro", "00-Intro\00-Intro.vcxproj", "{6849F7B7-30B7-4886-93C4-BE428757A111}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "05-SceneManager", "05-SceneManager\05-SceneManager.vcxproj", "{30B52332-21EE-43B1-91BF-B2494C04C34F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,20 +17,6 @@ Global
 		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{959C78E1-3BC0-4E30-BB42-048A1F251A85}.Debug|Any CPU.ActiveCfg = Debug|Win32
-		{959C78E1-3BC0-4E30-BB42-048A1F251A85}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
-		{959C78E1-3BC0-4E30-BB42-048A1F251A85}.Debug|Mixed Platforms.Build.0 = Debug|Win32
-		{959C78E1-3BC0-4E30-BB42-048A1F251A85}.Debug|Win32.ActiveCfg = Debug|Win32
-		{959C78E1-3BC0-4E30-BB42-048A1F251A85}.Debug|Win32.Build.0 = Debug|Win32
-		{959C78E1-3BC0-4E30-BB42-048A1F251A85}.Debug|x64.ActiveCfg = Debug|x64
-		{959C78E1-3BC0-4E30-BB42-048A1F251A85}.Debug|x64.Build.0 = Debug|x64
-		{959C78E1-3BC0-4E30-BB42-048A1F251A85}.Release|Any CPU.ActiveCfg = Release|Win32
-		{959C78E1-3BC0-4E30-BB42-048A1F251A85}.Release|Mixed Platforms.ActiveCfg = Release|Win32
-		{959C78E1-3BC0-4E30-BB42-048A1F251A85}.Release|Mixed Platforms.Build.0 = Release|Win32
-		{959C78E1-3BC0-4E30-BB42-048A1F251A85}.Release|Win32.ActiveCfg = Release|Win32
-		{959C78E1-3BC0-4E30-BB42-048A1F251A85}.Release|Win32.Build.0 = Release|Win32
-		{959C78E1-3BC0-4E30-BB42-048A1F251A85}.Release|x64.ActiveCfg = Release|x64
-		{959C78E1-3BC0-4E30-BB42-048A1F251A85}.Release|x64.Build.0 = Release|x64
 		{AADFB8E6-BDA3-4F69-9F0F-5C30EE6EACBA}.Debug|Any CPU.ActiveCfg = Debug|Win32
 		{AADFB8E6-BDA3-4F69-9F0F-5C30EE6EACBA}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
 		{AADFB8E6-BDA3-4F69-9F0F-5C30EE6EACBA}.Debug|Mixed Platforms.Build.0 = Debug|Win32
@@ -55,62 +31,6 @@ Global
 		{AADFB8E6-BDA3-4F69-9F0F-5C30EE6EACBA}.Release|Win32.Build.0 = Release|Win32
 		{AADFB8E6-BDA3-4F69-9F0F-5C30EE6EACBA}.Release|x64.ActiveCfg = Release|x64
 		{AADFB8E6-BDA3-4F69-9F0F-5C30EE6EACBA}.Release|x64.Build.0 = Release|x64
-		{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}.Debug|Any CPU.ActiveCfg = Debug|Win32
-		{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
-		{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}.Debug|Mixed Platforms.Build.0 = Debug|Win32
-		{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}.Debug|Win32.ActiveCfg = Debug|Win32
-		{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}.Debug|Win32.Build.0 = Debug|Win32
-		{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}.Debug|x64.ActiveCfg = Debug|x64
-		{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}.Debug|x64.Build.0 = Debug|x64
-		{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}.Release|Any CPU.ActiveCfg = Release|Win32
-		{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}.Release|Mixed Platforms.ActiveCfg = Release|Win32
-		{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}.Release|Mixed Platforms.Build.0 = Release|Win32
-		{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}.Release|Win32.ActiveCfg = Release|Win32
-		{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}.Release|Win32.Build.0 = Release|Win32
-		{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}.Release|x64.ActiveCfg = Release|x64
-		{5BB3DC44-D6AF-494B-9841-43B9AE3C6588}.Release|x64.Build.0 = Release|x64
-		{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}.Debug|Any CPU.ActiveCfg = Debug|Win32
-		{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
-		{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}.Debug|Mixed Platforms.Build.0 = Debug|Win32
-		{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}.Debug|Win32.ActiveCfg = Debug|Win32
-		{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}.Debug|Win32.Build.0 = Debug|Win32
-		{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}.Debug|x64.ActiveCfg = Debug|x64
-		{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}.Debug|x64.Build.0 = Debug|x64
-		{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}.Release|Any CPU.ActiveCfg = Release|Win32
-		{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}.Release|Mixed Platforms.ActiveCfg = Release|Win32
-		{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}.Release|Mixed Platforms.Build.0 = Release|Win32
-		{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}.Release|Win32.ActiveCfg = Release|Win32
-		{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}.Release|Win32.Build.0 = Release|Win32
-		{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}.Release|x64.ActiveCfg = Release|x64
-		{0CEDCE73-9775-46A9-9D9D-B8424E7D599D}.Release|x64.Build.0 = Release|x64
-		{6849F7B7-30B7-4886-93C4-BE428757A111}.Debug|Any CPU.ActiveCfg = Debug|Win32
-		{6849F7B7-30B7-4886-93C4-BE428757A111}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
-		{6849F7B7-30B7-4886-93C4-BE428757A111}.Debug|Mixed Platforms.Build.0 = Debug|Win32
-		{6849F7B7-30B7-4886-93C4-BE428757A111}.Debug|Win32.ActiveCfg = Debug|Win32
-		{6849F7B7-30B7-4886-93C4-BE428757A111}.Debug|Win32.Build.0 = Debug|Win32
-		{6849F7B7-30B7-4886-93C4-BE428757A111}.Debug|x64.ActiveCfg = Debug|x64
-		{6849F7B7-30B7-4886-93C4-BE428757A111}.Debug|x64.Build.0 = Debug|x64
-		{6849F7B7-30B7-4886-93C4-BE428757A111}.Release|Any CPU.ActiveCfg = Release|Win32
-		{6849F7B7-30B7-4886-93C4-BE428757A111}.Release|Mixed Platforms.ActiveCfg = Release|Win32
-		{6849F7B7-30B7-4886-93C4-BE428757A111}.Release|Mixed Platforms.Build.0 = Release|Win32
-		{6849F7B7-30B7-4886-93C4-BE428757A111}.Release|Win32.ActiveCfg = Release|Win32
-		{6849F7B7-30B7-4886-93C4-BE428757A111}.Release|Win32.Build.0 = Release|Win32
-		{6849F7B7-30B7-4886-93C4-BE428757A111}.Release|x64.ActiveCfg = Release|x64
-		{6849F7B7-30B7-4886-93C4-BE428757A111}.Release|x64.Build.0 = Release|x64
-		{30B52332-21EE-43B1-91BF-B2494C04C34F}.Debug|Any CPU.ActiveCfg = Debug|Win32
-		{30B52332-21EE-43B1-91BF-B2494C04C34F}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
-		{30B52332-21EE-43B1-91BF-B2494C04C34F}.Debug|Mixed Platforms.Build.0 = Debug|Win32
-		{30B52332-21EE-43B1-91BF-B2494C04C34F}.Debug|Win32.ActiveCfg = Debug|Win32
-		{30B52332-21EE-43B1-91BF-B2494C04C34F}.Debug|Win32.Build.0 = Debug|Win32
-		{30B52332-21EE-43B1-91BF-B2494C04C34F}.Debug|x64.ActiveCfg = Debug|x64
-		{30B52332-21EE-43B1-91BF-B2494C04C34F}.Debug|x64.Build.0 = Debug|x64
-		{30B52332-21EE-43B1-91BF-B2494C04C34F}.Release|Any CPU.ActiveCfg = Release|Win32
-		{30B52332-21EE-43B1-91BF-B2494C04C34F}.Release|Mixed Platforms.ActiveCfg = Release|Win32
-		{30B52332-21EE-43B1-91BF-B2494C04C34F}.Release|Mixed Platforms.Build.0 = Release|Win32
-		{30B52332-21EE-43B1-91BF-B2494C04C34F}.Release|Win32.ActiveCfg = Release|Win32
-		{30B52332-21EE-43B1-91BF-B2494C04C34F}.Release|Win32.Build.0 = Release|Win32
-		{30B52332-21EE-43B1-91BF-B2494C04C34F}.Release|x64.ActiveCfg = Release|x64
-		{30B52332-21EE-43B1-91BF-B2494C04C34F}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
